fix tooltip disappearance
In one very specific case, tooltips were being removed from the page
when new search results were rendered. This fixes that issue by stashing
a reference to the tooltip's "original parent" and returning it to that
element when the tooltip is hidden, so that if the element in which it
was shown is removed, the tooltip isn't removed with it.