render a subset of template directives on search
This started as an IE11 bug fix and ended as a performance optimization.
IE11 was throwing untraceable errors somewhere inside tagalong() with
the long list of template directives specified in
`picc.school.directives`. These are all used on the school page (where
they appear not to be a problem), but on search we only use a small
subset of them to render each result item, so we use a subset of the
directive keys. This improves performance and appears to resolve the
IE11 crashes when rendering new data after a change to search criteria.