clear results list on IE to avoid scary hard crash
This change introduces a bit of code that removes all of the elements
that get updated when the search results come back. Specifically, I
think that IE was crashing in cases when the number of items changed
(possibly only when it decreased), so to avoid the rudimentary "data
join" logic in tagalong, we just remove all of the children of the
results list before calling tagalong().

The important thing to note here is that **we only do this in IE *and*
if results have been loaded already**. In that case, tagalong has
already stashed a reference to the "template" node that gets used to
render each result item, so removing all of the children is safe. (This
is what happens when we get a zero-length result set anyway.)